# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 吉小农 (JLAU AI Assistant) API
  version: 1.0.0
  description: |
    吉农 AI 助手「吉小农」MVP 薄壳代理 + 轻量数据 API。
    核心对话经云函数代理转发 Coze 智能体，以 SSE (text/event-stream) 流式返回；
    其余为轻量 JSON 端点（反馈 / 转人工 / 场景注册表 / 猜你想问）。

    本文件是前后端联调的唯一契约（依据 SPEC.md §5）。所有端点路径必须带 /api/v1 版本前缀。

    两种 content-type 约定：
      - JSON 端点（feedback / human-handoff / scenarios / features）：
          响应头 Content-Type: application/json；统一结构 { code, data, message }。
      - 流式端点（chat）：
          响应头 Content-Type: text/event-stream；事件格式见 #/paths/~1api~1v1~1chat 描述与
          #/components/schemas 下的 *Event 定义。

    错误码（code 字段，0 表示成功）：
      0     成功
      4001  请求参数非法或缺失（invalid_request）
      4002  场景不存在（scenario_not_found）
      4003  反馈类型非法（invalid_feedback_type）
      4290  触发速率限制（rate_limited）
      5001  Coze 上游不可用（upstream_unavailable）
      5002  服务端内部错误（internal_error）
  contact:
    name: 吉农 AI 助手项目组
    url: https://github.com/jlau-ai-assistant

servers:
  - url: https://{env}.example.com
    description: 云函数环境
    variables:
      env:
        default: api
        enum:
          - dev
          - api

tags:
  - name: chat
    description: 对话流转发（SSE 流式）
  - name: feedback
    description: 答案反馈（有帮助 / 报错）
  - name: handoff
    description: 兜底转人工登记
  - name: registry
    description: 场景注册表与猜你想问（公开只读）

paths:
  /api/v1/chat:
    post:
      summary: 转发 Coze 对话（SSE 流式返回）
      description: |
        薄壳将用户消息经云函数代理转发至 Coze `POST /v3/chat`（stream:true），
        前端不持有 Coze token。响应为 SSE 流（text/event-stream），按事件顺序推送。

        SSE 事件格式（每条事件以 `event:` 行 + `data:` 行 + 空行分隔）：
          event: token      data: {"type":"token","content":"录取通知"}          # 增量文本 token（可多次）
          event: sources    data: {"type":"sources","items":[{...}]}             # RAG 来源标注（0 或 1 次）
          event: fallback   data: {"type":"fallback","guesses":[...],"contact":{...}} # 低置信/无答案引导
          event: done       data: {"type":"done","conversation_id":"...","message_id":"...","finish_reason":"stop"} # 流结束标志
          event: error      data: {"type":"error","code":5001,"message":"..."}   # 流中错误（出现即中止）

        结束标志：推送 `event: done` 后由服务端关闭连接；客户端以收到 `done` 事件为完成信号，
        不依赖固定超时。错误处理：若流建立前即失败，返回对应 HTTP 状态码 + JSON 错误体（见 400/429/500）。
      operationId: postChat
      tags:
        - chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              success:
                summary: 报到要带什么（成功流）
                value:
                  scenario_id: baodao
                  message: 报到要带什么
                  history: []
                  conversation_id: test-001
              fallback:
                summary: 知识库无答案（兜底流）
                value:
                  scenario_id: baodao
                  message: 吉农保研率多少
                  history: []
                  conversation_id: test-002
      responses:
        '200':
          description: SSE 流式响应（text/event-stream）
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/SSEStream'
              examples:
                success-stream:
                  summary: 成功流事件序列
                  value: |
                    event: token
                    data: {"type":"token","content":"新生报到需携带以下材料："}

                    event: token
                    data: {"type":"token","content":"1. 录取通知书；2. 身份证；3. 档案；4. 户口迁移证（如需）；5. 10 张小 2 寸照。"}

                    event: sources
                    data: {"type":"sources","items":[{"title":"2026 新生入学须知","url":"https://www.jlau.edu.cn/notice/2026","updated_at":"2026-07-20"}]}

                    event: done
                    data: {"type":"done","conversation_id":"test-001","message_id":"m-8f3a","finish_reason":"stop"}
                fallback-stream:
                  summary: 兜底流事件序列
                  value: |
                    event: fallback
                    data: {"type":"fallback","guesses":["报到流程是什么","宿舍怎么分配","学费怎么交"],"contact":{"name":"学工处","phone":"0431-84532980"}}

                    event: done
                    data: {"type":"done","conversation_id":"test-002","message_id":"m-2b7c","finish_reason":"no_answer"}
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'

  /api/v1/feedback:
    post:
      summary: 答案反馈（helpful / reported）
      description: |
        记录用户对单条回答的反馈事件，写入 feedback 集合，驱动知识库迭代 / 未知问题队列。
        响应为 JSON（Content-Type: application/json），统一结构 { code, data, message }。
      operationId: postFeedback
      tags:
        - feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
            examples:
              helpful:
                summary: 标记有帮助
                value:
                  message_id: m-8f3a
                  type: helpful
              reported:
                summary: 标记报错（附备注）
                value:
                  message_id: m-8f3a
                  type: reported
                  note: 来源链接已失效
      responses:
        '200':
          description: 反馈已记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseFeedback'
              examples:
                ok:
                  value:
                    code: 0
                    data:
                      id: fb-5c1d
                      message_id: m-8f3a
                      type: helpful
                    message: ok
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'

  /api/v1/human-handoff:
    post:
      summary: 兜底转人工登记
      description: |
        低置信 / 无答案 / 用户点「答错了」时登记转人工工单，写入 human_handoff 集合，进入兜底队列。
        contact 为可选字段，仅用于学工/辅导员回访；按 SPEC §10 合规约束，建议填微信昵称等粗粒度标识，
        不采集手机号、身份证、家庭地址等隐私字段。响应为 JSON，统一结构 { code, data, message }。
      operationId: postHumanHandoff
      tags:
        - handoff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HumanHandoffRequest'
            examples:
              with-contact:
                summary: 登记转人工（附联系标识）
                value:
                  scenario_id: baodao
                  question: 吉农保研率多少
                  contact: 微信：xiaoming2026
      responses:
        '200':
          description: 转人工工单已登记
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseHandoff'
              examples:
                ok:
                  value:
                    code: 0
                    data:
                      id: hh-9a2e
                      scenario_id: baodao
                      status: pending
                    message: ok
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'

  /api/v1/scenarios:
    get:
      summary: 分类入口列表（含 Lucide 图标名）
      description: |
        返回场景注册表（驱动分类入口与聊天主页标题）。MVP 仅 1 条（baodao 新生报到），
        其余进 Backlog。响应为 JSON，统一结构 { code, data:[...], message }。
        `lucide_icon` 为图标映射单一事实源（源自 scenarios 集合的 lucide_icon_name 字段），
        `icon` 为兼容别名，值与 lucide_icon 一致。
      operationId: getScenarios
      tags:
        - registry
      responses:
        '200':
          description: 场景列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseScenarios'
              examples:
                mvp:
                  summary: MVP 仅 baodao
                  value:
                    code: 0
                    data:
                      - id: baodao
                        name: 新生报到
                        icon: MapPin
                        lucide_icon: MapPin
                    message: ok
        '500':
          $ref: '#/components/responses/ServerError'

  /api/v1/features:
    get:
      summary: 快捷问题 / 猜你想问卡片
      description: |
        返回某场景的「猜你想问」卡片（features 集合）。可按 scenario_id 查询参数过滤。
        响应为 JSON，统一结构 { code, data:[...], message }。MVP 写入报到域 4~6 条。
      operationId: getFeatures
      tags:
        - registry
      parameters:
        - name: scenario_id
          in: query
          description: 场景过滤；缺省返回全部场景的快捷卡
          required: false
          schema:
            type: string
            example: baodao
      responses:
        '200':
          description: 猜你想问卡片列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseFeatures'
              examples:
                baodao:
                  summary: 报到域猜你想问
                  value:
                    code: 0
                    data:
                      - id: f1
                        text: 报到要带什么材料
                        scenario_id: baodao
                      - id: f2
                        text: 从长春站怎么去学校
                        scenario_id: baodao
                      - id: f3
                        text: 宿舍怎么分配
                        scenario_id: baodao
                      - id: f4
                        text: 学费怎么交
                        scenario_id: baodao
                      - id: f5
                        text: 校园卡怎么激活
                        scenario_id: baodao
                      - id: f6
                        text: 军训要准备什么
                        scenario_id: baodao
                    message: ok
        '500':
          $ref: '#/components/responses/ServerError'

  # ================= 反馈管理后台（方案 A，A-1~A-4） =================
  # 除 login 外均需请求头 X-Admin-Token: <av1.*>；未配置 ADMIN_PASSWORD 时一律 503 code 4011。

  /api/v1/admin/login:
    post:
      summary: 管理员登录（校验 ADMIN_PASSWORD，签发会话令牌）
      description: |
        匿名可调（自身带登录失败限流：同 IP 15min 失败 ≥5 次 → 429 code 4290 + Retry-After）。
        未配置 ADMIN_PASSWORD → 503 code 4011「管理后台未启用」，无空密码放行路径。
        响应 data.token 为 av1 前缀 HMAC 会话令牌，随后续请求经 X-Admin-Token 头携带。
      operationId: adminLogin
      tags:
        - admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  description: ADMIN_PASSWORD 环境变量值
                  example: s3cret
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAdminLogin'
              examples:
                ok:
                  value:
                    code: 0
                    data:
                      token: av1.1789000000000.nQ3fAbCdEfGh.abcdefghijklmnopqrstuvwxyz0123456789
                      expires_at: 1789000000000
                    message: ok
        '401':
          description: 密码错误（code=4013）
          $ref: '#/components/responses/AdminUnauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/AdminDisabled'

  /api/v1/admin/logout:
    post:
      summary: 管理员登出（轻量对称端点）
      description: |
        需有效管理员令牌。服务端不维护全局黑名单（靠 TTL），退出 = 前端清除本地令牌；
        本端点仅返回 { ok: true } 供客户端对称调用。
      operationId: adminLogout
      tags:
        - admin
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - data
                  - message
                properties:
                  code:
                    type: integer
                    example: 0
                  data:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: true
                  message:
                    type: string
                    example: ok
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '503':
          $ref: '#/components/responses/AdminDisabled'

  /api/v1/admin/feedback:
    get:
      summary: 反馈列表（倒序 + 筛选 + 分页）
      description: |
        需有效管理员令牌。支持 type / scenario_id / q（note 模糊匹配，不区分大小写）/
        page / pageSize（默认 20）。返回 items 含 `_id`（两 store 同构），
        created_at 倒序（新→旧）。旧记录无 snapshot 字段（升级前数据）。
      operationId: adminListFeedback
      tags:
        - admin
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [helpful, reported]
          description: 反馈类型筛选
        - name: scenario_id
          in: query
          schema:
            type: string
          description: 场景标识筛选
        - name: q
          in: query
          schema:
            type: string
          description: note 模糊搜索
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 页码（1-based）
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            default: 20
          description: 每页条数
      responses:
        '200':
          description: 反馈列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAdminFeedbackList'
              examples:
                ok:
                  value:
                    code: 0
                    data:
                      items:
                        - _id: fb-5c1d
                          message_id: m-8f3a
                          type: reported
                          note: 缴费入口写错了
                          scenario_id: baodao
                          created_at: '2026-08-07T02:30:00.000Z'
                          snapshot:
                            question: 报到要交多少钱？
                            answer: 学费需在报到时现场缴纳。
                      total: 1
                      page: 1
                      pageSize: 20
                    message: ok
        '400':
          description: type 非法（code=4001）
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '503':
          $ref: '#/components/responses/AdminDisabled'

  /api/v1/admin/kb/refresh:
    post:
      summary: 知识库热刷新（清缓存并强制重读磁盘）
      description: |
        需有效管理员令牌。清空 server/src/kb/retrieve.ts 模块缓存并立即重读一次；
        成功返回 { cleared: true, doc_count }；目录不存在 / 读取失败返回
        500 code 5002 明确文案（不静默）。
      operationId: adminRefreshKb
      tags:
        - admin
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAdminKbRefresh'
              examples:
                ok:
                  value:
                    code: 0
                    data:
                      cleared: true
                      doc_count: 12
                    message: ok
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/AdminDisabled'

components:
  schemas:
    # ---------- 请求体 ----------
    ChatMessage:
      type: object
      description: 历史消息单元（多轮上下文）
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum:
            - user
            - assistant
          description: 发言角色
        content:
          type: string
          description: 消息文本
          example: 从长春站怎么去学校

    ChatRequest:
      type: object
      required:
        - scenario_id
        - message
        - conversation_id
      properties:
        scenario_id:
          type: string
          description: 场景标识（须为 scenarios 注册表中的 id）
          example: baodao
        message:
          type: string
          description: 当前用户提问
          minLength: 1
          maxLength: 2000
          example: 报到要带什么
        history:
          type: array
          description: 多轮上下文（不含当前消息）；数组为空表示首轮
          items:
            $ref: '#/components/schemas/ChatMessage'
          default: []
        conversation_id:
          type: string
          description: 会话标识，用于 Coze conversation 续接与消息关联
          example: test-001

    FeedbackRequest:
      type: object
      required:
        - message_id
        - type
      properties:
        message_id:
          type: string
          description: 被反馈的回答 message_id（来自 chat 流 done 事件）
          example: m-8f3a
        type:
          type: string
          enum:
            - helpful
            - reported
          description: helpful=有帮助；reported=报错
          example: helpful
        note:
          type: string
          nullable: true
          description: 报错补充说明（可选）
          maxLength: 500
          example: 来源链接已失效
        snapshot:
          type: object
          nullable: true
          description: |
            问答快照（方案 A A-2，可选）。仅「报错」时由前端携带：
            该条 AI 回答 + 其前一条用户提问。不传/为 null 时行为与旧版一致。
          properties:
            question:
              type: string
              description: 用户提问（≤2000 字符，超限返 400 code 4004）
              maxLength: 2000
              example: 报到要交多少钱？
            answer:
              type: string
              description: AI 回答（≤20000 字符，超限返 400 code 4004）
              maxLength: 20000
              example: 学费需在报到时现场缴纳。
          required:
            - question
            - answer

    HumanHandoffRequest:
      type: object
      required:
        - scenario_id
        - question
      properties:
        scenario_id:
          type: string
          description: 触发转人工的场景标识
          example: baodao
        question:
          type: string
          description: 用户原始问题（脱敏摘要，不存完整原始输入全文）
          minLength: 1
          maxLength: 2000
          example: 吉农保研率多少
        contact:
          type: string
          nullable: true
          description: 可选回访标识（建议微信昵称等粗粒度，不采集手机号/身份证/家庭地址）
          maxLength: 100
          example: 微信：xiaoming2026

    # ---------- SSE 事件 ----------
    SSEStream:
      type: string
      description: |
        SSE 文本流。每条事件形如：
          event: <EventType>\n
          data: <JSON>\n
          \n
        事件类型与其 data 载荷见下方 *Event 定义。结束标志为 event: done 后连接关闭。
      format: binary

    TokenEvent:
      type: object
      description: 增量文本 token 事件（data 载荷）
      required:
        - type
        - content
      properties:
        type:
          type: string
          enum: [token]
        content:
          type: string
          description: 本次推送的文本片段
          example: 新生报到需携带以下材料：

    SourcesEvent:
      type: object
      description: RAG 来源标注事件（data 载荷）
      required:
        - type
        - items
      properties:
        type:
          type: string
          enum: [sources]
        items:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
                example: 2026 新生入学须知
              url:
                type: string
                format: uri
                example: https://www.jlau.edu.cn/notice/2026
              updated_at:
                type: string
                format: date
                example: '2026-07-20'

    FallbackEvent:
      type: object
      description: 低置信 / 无答案兜底引导事件（data 载荷）
      required:
        - type
        - guesses
        - contact
      properties:
        type:
          type: string
          enum: [fallback]
        guesses:
          type: array
          description: 猜你想问（引导用户继续）
          items:
            type: string
          example:
            - 报到流程是什么
            - 宿舍怎么分配
            - 学费怎么交
        contact:
          type: object
          description: 转人工联系入口
          properties:
            name:
              type: string
              example: 学工处
            phone:
              type: string
              example: '0431-84532980'

    DoneEvent:
      type: object
      description: 流结束标志事件（data 载荷），出现后连接关闭
      required:
        - type
        - conversation_id
        - message_id
        - finish_reason
      properties:
        type:
          type: string
          enum: [done]
        conversation_id:
          type: string
          example: test-001
        message_id:
          type: string
          description: 本回答唯一标识，供 feedback 引用
          example: m-8f3a
        finish_reason:
          type: string
          enum:
            - stop
            - no_answer
          description: stop=正常结束；no_answer=无答案走兜底
          example: stop

    SSEErrorEvent:
      type: object
      description: 流中错误事件（data 载荷），出现即中止后续推送
      required:
        - type
        - code
        - message
      properties:
        type:
          type: string
          enum: [error]
        code:
          type: integer
          example: 5001
        message:
          type: string
          example: 对话服务暂不可用

    # ---------- 公共响应结构 ----------
    ScenarioItem:
      type: object
      required:
        - id
        - name
        - icon
        - lucide_icon
      properties:
        id:
          type: string
          example: baodao
        name:
          type: string
          example: 新生报到
        icon:
          type: string
          description: 兼容别名，值与 lucide_icon 一致
          example: MapPin
        lucide_icon:
          type: string
          description: Lucide 图标名（图标映射单一事实源，源自 scenarios.lucide_icon_name）
          example: MapPin

    FeatureItem:
      type: object
      required:
        - id
        - text
        - scenario_id
      properties:
        id:
          type: string
          example: f1
        text:
          type: string
          example: 报到要带什么材料
        scenario_id:
          type: string
          example: baodao

    ApiResponseFeedback:
      type: object
      description: 统一响应结构（feedback）
      required:
        - code
        - data
        - message
      properties:
        code:
          type: integer
          example: 0
        data:
          type: object
          nullable: true
          properties:
            id:
              type: string
              example: fb-5c1d
            message_id:
              type: string
              example: m-8f3a
            type:
              type: string
              enum:
                - helpful
                - reported
        message:
          type: string
          example: ok

    ApiResponseAdminLogin:
      type: object
      description: 统一响应结构（admin login）
      required:
        - code
        - data
        - message
      properties:
        code:
          type: integer
          example: 0
        data:
          type: object
          nullable: true
          properties:
            token:
              type: string
              description: av1 前缀 HMAC 会话令牌（请求头 X-Admin-Token 携带）
              example: av1.1789000000000.nQ3fAbCdEfGh.abcdefghijklmnopqrstuvwxyz0123456789
            expires_at:
              type: integer
              description: 过期时间戳（毫秒）
              example: 1789000000000
        message:
          type: string
          example: ok

    AdminFeedbackItem:
      type: object
      description: 管理端反馈列表项（两 store 同构，必含 _id）
      required:
        - _id
        - message_id
        - type
        - note
        - scenario_id
        - created_at
      properties:
        _id:
          type: string
          example: fb-5c1d
        message_id:
          type: string
          example: m-8f3a
        type:
          type: string
          enum:
            - helpful
            - reported
        note:
          type: string
          nullable: true
          example: 缴费入口写错了
        scenario_id:
          type: string
          example: baodao
        created_at:
          type: string
          format: date-time
          description: ISO-8601 UTC
          example: '2026-08-07T02:30:00.000Z'
        snapshot:
          type: object
          nullable: true
          description: 问答快照；旧记录（升级前）缺失该字段
          properties:
            question:
              type: string
              example: 报到要交多少钱？
            answer:
              type: string
              example: 学费需在报到时现场缴纳。

    ApiResponseAdminFeedbackList:
      type: object
      description: 统一响应结构（admin 反馈列表）
      required:
        - code
        - data
        - message
      properties:
        code:
          type: integer
          example: 0
        data:
          type: object
          nullable: true
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AdminFeedbackItem'
            total:
              type: integer
              example: 1
            page:
              type: integer
              example: 1
            pageSize:
              type: integer
              example: 20
        message:
          type: string
          example: ok

    ApiResponseAdminKbRefresh:
      type: object
      description: 统一响应结构（admin kb refresh）
      required:
        - code
        - data
        - message
      properties:
        code:
          type: integer
          example: 0
        data:
          type: object
          nullable: true
          properties:
            cleared:
              type: boolean
              example: true
            doc_count:
              type: integer
              example: 12
        message:
          type: string
          example: ok

    ApiResponseHandoff:
      type: object
      description: 统一响应结构（human-handoff）
      required:
        - code
        - data
        - message
      properties:
        code:
          type: integer
          example: 0
        data:
          type: object
          nullable: true
          properties:
            id:
              type: string
              example: hh-9a2e
            scenario_id:
              type: string
              example: baodao
            status:
              type: string
              enum:
                - pending
                - contacted
                - resolved
        message:
          type: string
          example: ok

    ApiResponseScenarios:
      type: object
      description: 统一响应结构（scenarios）
      required:
        - code
        - data
        - message
      properties:
        code:
          type: integer
          example: 0
        data:
          type: array
          items:
            $ref: '#/components/schemas/ScenarioItem'
        message:
          type: string
          example: ok

    ApiResponseFeatures:
      type: object
      description: 统一响应结构（features）
      required:
        - code
        - data
        - message
      properties:
        code:
          type: integer
          example: 0
        data:
          type: array
          items:
            $ref: '#/components/schemas/FeatureItem'
        message:
          type: string
          example: ok

  responses:
    BadRequest:
      description: 请求参数非法或缺失（code=4001/4002/4003/4004）
      content:
        application/json:
          schema:
            type: object
            required:
              - code
              - data
              - message
            properties:
              code:
                type: integer
                enum:
                  - 4001
                  - 4002
                  - 4003
                  - 4004
              data:
                type: object
                nullable: true
              message:
                type: string
                example: 缺少必填字段 scenario_id
    AdminUnauthorized:
      description: 未登录 / 管理员令牌无效或过期（code=4012）；密码错误（code=4013）
      content:
        application/json:
          schema:
            type: object
            required:
              - code
              - data
              - message
            properties:
              code:
                type: integer
                enum:
                  - 4012
                  - 4013
              data:
                type: object
                nullable: true
              message:
                type: string
                example: 登录已失效，请重新登录
    AdminDisabled:
      description: 管理后台未启用（未配置 ADMIN_PASSWORD，code=4011）
      content:
        application/json:
          schema:
            type: object
            required:
              - code
              - data
              - message
            properties:
              code:
                type: integer
                enum:
                  - 4011
              data:
                type: object
                nullable: true
              message:
                type: string
                example: 管理后台未启用
    TooManyRequests:
      description: 触发速率限制（code=4290）
      content:
        application/json:
          schema:
            type: object
            required:
              - code
              - data
              - message
            properties:
              code:
                type: integer
                enum:
                  - 4290
              data:
                type: object
                nullable: true
              message:
                type: string
                example: 请求过于频繁，请稍后再试
    ServerError:
      description: 服务端 / 上游错误（code=5001/5002）
      content:
        application/json:
          schema:
            type: object
            required:
              - code
              - data
              - message
            properties:
              code:
                type: integer
                enum:
                  - 5001
                  - 5002
              data:
                type: object
                nullable: true
              message:
                type: string
                example: 对话服务暂不可用